Avoid using a variable name that is the same as the package name
The current example codes use `url` for a variable name. This is
the same as the package name. So after the definition of the
variable, features provide by the package don't work.

I think this is a common mistake, but it should be avoided in
examples because it can be confusing.